Alle Seiten haben jetzt den gleichen Drawer (zentralisiert) Hardcoded Strings sind jetzt in Strings.xml
