+++ AndroidMockup/Mockup2/app/src/main/res/layout/activity_base.xml
@@ -9,7 +9,7 @@
     tools:openDrawer="start">
 
     <include
-        layout="@layout/app_bar_freundesliste"
+        layout="@layout/app_bar_base"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
