+++ AndroidMockup/Mockup2/app/src/main/res/layout/activity_main.xml
@@ -19,7 +19,7 @@
         android:layout_height="match_parent"
         android:layout_gravity="start"
         android:fitsSystemWindows="true"
-        app:headerLayout="@layout/nav_header_main"
-        app:menu="@menu/activity_main_drawer" />
+        app:headerLayout="@layout/nav_header_base"
+        app:menu="@menu/activity_base_drawer" />
 
 </android.support.v4.widget.DrawerLayout>
