+++ AndroidMockup/Mockup2/app/src/main/res/values/strings.xml
@@ -80,4 +80,15 @@
     <string name="pref_title_vibrate">Vibrate</string>
     <string name="title_activity_freunde">Freunde</string>
     <string name="title_activity_freundesliste">Freundesliste</string>
+    <string name="title_activity_base">BaseActivity</string>
+    <string name="user_name">Thomas Siepmann</string>
+    <string name="thomas_siepman_fhaachen_de">thomas.siepmann@fhaachen.de</string>
+    <string name="user_email">siepmann</string>
+    <string name="home">Home</string>
+    <string name="rooms">Räume</string>
+    <string name="friendlist">Freundesliste</string>
+    <string name="settings">Einstellungen</string>
+    <string name="getMaster">Professor</string>
+    <string name="timetable">Stundenplan</string>
+    <string name="profile_pic">Profilbild</string>
 </resources>
